absorb canvas deserializer
